requirements.txt

Django==3.0.3
django-cors-headers==3.2.1
whitenoise==5.2.0
psycopg2==2.8.6
dj-database-url==0.5.0
gunicorn==20.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
django-import-export